[fix] Always set the run name when initializing a run in lightning (#3102)
Always set the run name when initializing a run in lightning

Before, the run name would only be set when a run hash has been specified by the user, which
is only the case in unusual circumstances, e.g. when a previous model training is resumed
at a later time. In effect, run names were never set in my case.